+++ AlixSystemLoader/src/main/java/alix/common/antibot/firewall/ataraxia/AlixAtaraxia.java
@@ -3,6 +3,7 @@ package alix.common.antibot.firewall.ataraxia;
 import io.netty.channel.epoll.Epoll;
 
 import java.net.InetAddress;
+import java.util.List;
 
 public final class AlixAtaraxia {
 
@@ -11,18 +12,19 @@ public final class AlixAtaraxia {
     }
 
     public static void blacklist(InetAddress ip) {
+        AtaraxiaIPC.mapUpdate_writeAndFlush(true, true, List.of(ip));
     }
 
     public static void whitelist(InetAddress ip) {
-
+        AtaraxiaIPC.mapUpdate_writeAndFlush(false, true, List.of(ip));
     }
 
-    public static void unblacklist(String ip) {
-
+    public static void unblacklist(InetAddress ip) {
+        AtaraxiaIPC.mapUpdate_writeAndFlush(true, false, List.of(ip));
     }
 
     public static void unwhitelist(InetAddress ip) {
-
+        AtaraxiaIPC.mapUpdate_writeAndFlush(false, false, List.of(ip));
     }
 
     public static boolean isEnabled() {