+++ AlixSystemLoader/src/main/java/alix/common/antibot/firewall/ataraxia/ConnectionState.java
@@ -1,0 +1,7 @@
+package alix.common.antibot.firewall.ataraxia;
+
+public enum ConnectionState {
+
+    DISCONNECTED, PRE_HANDSHAKE, POST_HANDSHAKE//, SYNC_COMPLETE
+
+}