+++ AlixSystemLoader/src/main/java/alix/common/connection/ratelimit/ConnectionEntry.java
@@ -1,0 +1,4 @@
+package alix.common.connection.ratelimit;
+
+record ConnectionEntry(long expiresAt) {
+}